refactor(nms): move fbc lib auth
