+++ nms/fbc_js_core/platform_server/host/middleware.js
@@ -15,7 +15,7 @@
  */
 
 import type {ExpressResponse, NextFunction} from 'express';
-import type {FBCNMSRequest} from '../../../fbc_js_core/auth/access';
+import type {FBCNMSRequest} from '../../../server/auth/access';
 
 import asyncHandler from '../../../fbc_js_core/util/asyncHandler';
