+++ nms/fbc_js_core/ui/components/auth/EditUserDialog.js
@@ -31,7 +31,7 @@ import React, {useCallback, useContext, useMemo, useState} from 'react';
 import Select from '@material-ui/core/Select';
 import TextField from '@material-ui/core/TextField';
 import renderList from '../../../../fbc_js_core/util/renderList';
-import {UserRoles} from '../../../../fbc_js_core/auth/types';
+import {UserRoles} from '../../../../shared/types';
 import {makeStyles} from '@material-ui/styles';
 
 export type EditUser = {
