+++ nms/server/auth/expressOnboarding.js
@@ -19,7 +19,7 @@ import type {FBCNMSRequest} from './access';
 
 import express from 'express';
 import staticDist from '../../fbc_js_core/webpack_config/staticDist';
-import {AccessRoles} from './roles';
+import {AccessRoles} from '../../shared/roles';
 import {Organization, User} from '../../fbc_js_core/sequelize_models';
 import {getPropsToUpdate} from './util';
