+++ nms/shared/errorCodes.js
