+++ nms/shared/types.js
